+++ app/src/main/java/com/example/whatstheweather/Models/Wind.java
@@ -2,7 +2,7 @@ package com.example.whatstheweather.Models;
 
 public class Wind {
     public double speed;
-    public int deg;
+    public float deg;
     public double gust;
 
     public  Wind(){
@@ -17,11 +17,11 @@ public class Wind {
         this.speed = speed;
     }
 
-    public int getDeg() {
+    public float getDeg() {
         return deg;
     }
 
-    public void setDeg(int deg) {
+    public void setDeg(float deg) {
         this.deg = deg;
     }
